refactored application directory, appended non-prod files to .gitignore
