+++ .gitattributes
@@ -1,4 +1,4 @@
-model.h5 filter=lfs diff=lfs merge=lfs -text
 glove.6B.100d.txt filter=lfs diff=lfs merge=lfs -text
+model.h5 filter=lfs diff=lfs merge=lfs -text
 sentiment140_trimmed.csv filter=lfs diff=lfs merge=lfs -text
 tokens.json filter=lfs diff=lfs merge=lfs -text
