--- glove.6B.100d.txt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e26d49a1aeb1aaf54ac117da4d013645ece82bb59ddd9ed8a74887b31e3683ca
-size 347516733
