+++ requirements.txt
@@ -1,13 +1,12 @@
-Flask==2.2.3
+Flask==2.2.2
 Flask_Caching==2.0.2
 Flask_Login==0.6.2
-flask_mysqldb==1.0.1
 flask_sqlalchemy==3.0.3
-keras==2.11.0
-Keras_Preprocessing==1.1.2
-numpy==1.24.2
+keras==2.12.0
+keras_preprocessing==1.1.2
+numpy==1.23.0
 pandas==1.5.3
 python-dotenv==1.0.0
-requests==2.28.2
+requests==2.28.1
 scikit_learn==1.2.2
-Werkzeug==2.2.2
+Werkzeug==2.2.3
