--- sentiment140_trimmed.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5c0026655353c89a2a548a63f707a04abf3165e44f7637bd2e894510cea5e4d2
-size 125819229
